refactor: reorganize components into design system
