+++ app/admin/pedidos/page.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import { useAuthGuard } from "@/lib/hooks/useAuthGuard";
 import { Pedido } from "@/types";
 import { saveAs } from "file-saver";
-import LoadingOverlay from "@/components/LoadingOverlay";
+import LoadingOverlay from "@/components/organisms/LoadingOverlay";
 import ModalEditarPedido from "./componentes/ModalEditarPedido";
 import { useToast } from "@/lib/context/ToastContext";
