+++ app/admin/redefinir-senha/page.tsx
@@ -2,7 +2,7 @@
 
 import dynamic from "next/dynamic";
 import { Suspense } from "react";
-import LoadingOverlay from "@/components/LoadingOverlay";
+import LoadingOverlay from "@/components/organisms/LoadingOverlay";
 
 // Ajuste o path aqui conforme a localização real do arquivo:
 const RedefinirSenhaClient = dynamic(
