+++ components/atoms/index.ts
@@ -1,0 +1,4 @@
+export { Button } from './Button';
+export { TextField } from './TextField';
+export { default as Spinner } from './Spinner';
+export { default as ToggleSwitch } from './ToggleSwitch';
