+++ components/molecules/SmoothTabs.tsx
