+++ components/molecules/index.ts
@@ -1,0 +1,4 @@
+export { FormField } from './FormField';
+export { InputWithMask } from './InputWithMask';
+export { default as SmoothTabs } from './SmoothTabs';
+export { default as SaldoCard } from './SaldoCard';
