+++ components/organisms/index.ts
@@ -1,0 +1,2 @@
+export { default as LoadingOverlay } from './LoadingOverlay';
+export { default as ModalAnimated } from './ModalAnimated';
