+++ docs/design-system.md
@@ -54,7 +54,7 @@ Os principais componentes têm exemplos interativos no Storybook. Alguns dos mai
 - Cabeçalhos (`stories/Header.stories.ts`)
 - Modais (`stories/ModalEditarPerfil.stories.tsx` e outros)
 - Tabs suaves (`stories/Tabs.stories.tsx`)
-- Overlays de loading (`components/LoadingOverlay.tsx`)
+ - Overlays de loading (`components/organisms/LoadingOverlay.tsx`)
 
 Execute `npm run storybook` para iniciar a interface e explorar os exemplos.
